move computeMillisInCurrentDay to common functionality
