+++ lib/carelink/carelinkSimulator.js
@@ -1,15 +1,15 @@
 /*
  * == BSD2 LICENSE ==
  * Copyright (c) 2014, Tidepool Project
- * 
+ *
  * This program is free software; you can redistribute it and/or modify it under
  * the terms of the associated License, which is identical to the BSD 2-Clause
  * License as published by the Open Source Initiative at opensource.org.
- * 
+ *
  * This program is distributed in the hope that it will be useful, but WITHOUT
  * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
  * FOR A PARTICULAR PURPOSE. See the License for more details.
- * 
+ *
  * You should have received a copy of the License along with this program; if
  * not, you can obtain one from Tidepool Project at tidepool.org.
  * == BSD2 LICENSE ==
@@ -20,6 +20,7 @@ var sundial = require('sundial');
 var util = require('util');
 
 var builder = require('./../objectBuilder.js');
+var common = require('../commonFunctions');
 
 
 var twentyFourHours = 24 * 60 * 60 * 1000;
@@ -57,24 +58,7 @@ function annotateEvent(event, ann){
   return event;
 }
 
-/**
- * Computes the number of milliseconds after midnight on the date specified.
- *
- * @param dateTime DateTime object to figure out millis from
- * @returns {number} number of millis in current day
- */
-function computeMillisInCurrentDay(e){
-  var msFromMidnight = sundial.getMsFromMidnight(e.time, e.timezoneOffset);
-  var fifteenMinsInMs = 15*60*1000;
-  // adjustments for clock drift screw up our logic to see if a basal
-  // matches a schedule, so we round to the nearest fifteen mins
-  // to increase the chance of matching up with a schedule
-  if (e.conversionOffset && e.conversionOffset !== 0) {
-    var result = Math.round(msFromMidnight/fifteenMinsInMs)*fifteenMinsInMs;
-    return result === 864e5 ? 0 : result;
-  }
-  return msFromMidnight;
-}
+
 
 
 /**
@@ -218,7 +202,7 @@ exports.make = function(config){
               if (currSchedule == null || currSchedule.length === 0) {
                 parts.splice(doneIndex, 1);
               } else {
-                var millisInDay = computeMillisInCurrentDay({
+                var millisInDay = common.computeMillisInCurrentDay({
                   time: doneTime,
                   timezoneOffset: sched.timezoneOffset,
                   conversionOffset: sched.conversionOffset
@@ -404,7 +388,7 @@ exports.make = function(config){
           return;
         }
       } else {
-        var millisInDay = computeMillisInCurrentDay(event);
+        var millisInDay = common.computeMillisInCurrentDay(event);
         for (var i = currSchedule.length - 1; i >= 0; --i) {
           if (currSchedule[i].start <= millisInDay) {
             break;
@@ -628,12 +612,12 @@ exports.make = function(config){
       if (currSettings != null) {
         var currSchedule = currSettings.basalSchedules[currSettings.activeSchedule];
         if (currSchedule.length > 0) {
-          var millisInDay = computeMillisInCurrentDay(event);
+          var millisInDay = common.computeMillisInCurrentDay(event);
           for (var i = currSchedule.length - 1; i >= 0; --i) {
             if (currSchedule[i].start <= millisInDay) {
               break;
             }
-          } 
+          }
 
           var basalBuilder = bob.makeScheduledBasal()
             .with_scheduleName(currSettings.activeSchedule)
@@ -883,7 +867,7 @@ exports.make = function(config){
           return;
         }
 
-        var millisInDay = computeMillisInCurrentDay(currSettings);
+        var millisInDay = common.computeMillisInCurrentDay(currSettings);
         for (var i = currSchedule.length - 1; i > 0; --i) {
           if (currSchedule[i].start <= millisInDay) {
             break;
