+++ lib/commonFunctions.js
@@ -19,6 +19,25 @@ var sundial = require('sundial');
 
 var annotate = require('./eventAnnotations');
 
+/**
+ * Computes the number of milliseconds after midnight on the date specified.
+ *
+ * @param dateTime DateTime object to figure out millis from
+ * @returns {number} number of millis in current day
+ */
+exports.computeMillisInCurrentDay = function(e){
+  var msFromMidnight = sundial.getMsFromMidnight(e.time, e.timezoneOffset);
+  var fifteenMinsInMs = 15*60*1000;
+  // adjustments for clock drift screw up our logic to see if a basal
+  // matches a schedule, so we round to the nearest fifteen mins
+  // to increase the chance of matching up with a schedule
+  if (e.conversionOffset && e.conversionOffset !== 0) {
+    var result = Math.round(msFromMidnight/fifteenMinsInMs)*fifteenMinsInMs;
+    return result === 864e5 ? 0 : result;
+  }
+  return msFromMidnight;
+};
+
 exports.bytes2hex = function(bytes, noGaps) {
   var message = '';
   for(var i in bytes) {
